requirements.txt
pip==23.0.1
Flask==2.2.2
gunicorn
requests>=2.28.2
urllib3<1.27,>=1.25.4 
python-dotenv
heyoo
psycopg2-binary==2.9.6
boto3
langchain
openai
google-search-results
pinecone-client
colorama==0.4.6
singleton
stripe
flask-cors
